requirements.txt
# Django
Django>=5.0,<6.0

# Django REST Framework
djangorestframework>=3.15.0

# Database
psycopg2-binary>=2.9.9

# Authentication
djangorestframework-simplejwt>=5.3.1

# Environment variables
python-decouple>=3.8

# CORS
django-cors-headers>=4.3.1

# Celery
celery>=5.3.4
redis>=5.0.1

# Django Channels (for WebSocket)
channels>=4.0.0
channels-redis>=4.2.0

# Password hashing
bcrypt>=4.1.2

# API Documentation
drf-spectacular>=0.27.0

# Monitoring
sentry-sdk>=2.0.0

# Email
django-anymail>=10.2

# Date/Time utilities
python-dateutil>=2.8.2

# Utilities
Pillow>=10.2.0

# Logging
python-json-logger>=2.0.7

